backend/requirements.txt
fastapi>=0.135.2
uvicorn[standard]>=0.35.0
pydantic>=2.12.5
pydantic-settings>=2.10.1
python-multipart>=0.0.20
Pillow>=11.3.0
ultralytics>=8.3.0
transformers>=4.36.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
python-dotenv>=1.0.1
tqdm>=4.66.0
motor>=3.7.1
pymongo>=4.13.2
PyJWT>=2.10.1
bcrypt>=4.3.0
email-validator>=2.3.0
pandas>=2.3.3
scikit-learn>=1.7.2
joblib>=1.5.2
yfinance>=0.2.66
